links from names to ala. reduced radius size for nearby occurrences
git-svn-id: http://ala-soils2sat.googlecode.com/svn/trunk/ala-soils2sat/trunk@73 852cf724-85a4-5791-e48c-ddb37d8467d3

+++ grails-app/controllers/au/org/ala/soils2sat/StudyLocationController.groovy
@@ -485,7 +485,7 @@ class StudyLocationController {
         def studyLocationSummary = studyLocationService.getStudyLocationSummary(studyLocationName)
 
         def rank = params.rank ?: 'family';
-        int radius = params.int("radius") ?: 10
+        float radius = params.float("radius") ?: 10
 
         def studyLocationTaxaList = biocacheService.getTaxaNamesForLocation(studyLocationSummary.latitude, studyLocationSummary.longitude, radius, rank)
